// test/integration/test/sql-orm-client/fixtures/polymorphism/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Account: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Admin: { readonly role: CodecTypes['pg/text@1']['output'] };
    readonly Bug: {
      readonly assigneeId: CodecTypes['pg/int4@1']['output'] | null;
      readonly severity: CodecTypes['pg/text@1']['output'];
    };
    readonly Epic: { readonly scope: CodecTypes['pg/text@1']['output'] };
    readonly Feature: {
      readonly assigneeId: CodecTypes['pg/int4@1']['output'] | null;
      readonly priority: CodecTypes['pg/int4@1']['output'];
    };
    readonly Person: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Project: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Regular: { readonly plan: CodecTypes['pg/text@1']['output'] | null };
    readonly Task: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly projectId: CodecTypes['pg/int4@1']['output'] | null;
      readonly reporterId: CodecTypes['pg/int4@1']['output'] | null;
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly type: CodecTypes['pg/text@1']['output'];
    };
    readonly TaskComment: {
      readonly body: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly taskId: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly Ticket: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly ownerId: CodecTypes['pg/int4@1']['output'] | null;
      readonly subject: CodecTypes['pg/text@1']['output'];
    };
    readonly User: {
      readonly accountId: CodecTypes['pg/int4@1']['output'] | null;
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly kind: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Account: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Admin: { readonly role: CodecTypes['pg/text@1']['input'] };
    readonly Bug: {
      readonly assigneeId: CodecTypes['pg/int4@1']['input'] | null;
      readonly severity: CodecTypes['pg/text@1']['input'];
    };
    readonly Epic: { readonly scope: CodecTypes['pg/text@1']['input'] };
    readonly Feature: {
      readonly assigneeId: CodecTypes['pg/int4@1']['input'] | null;
      readonly priority: CodecTypes['pg/int4@1']['input'];
    };
    readonly Person: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Project: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Regular: { readonly plan: CodecTypes['pg/text@1']['input'] | null };
    readonly Task: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly projectId: CodecTypes['pg/int4@1']['input'] | null;
      readonly reporterId: CodecTypes['pg/int4@1']['input'] | null;
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly type: CodecTypes['pg/text@1']['input'];
    };
    readonly TaskComment: {
      readonly body: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly taskId: CodecTypes['pg/int4@1']['input'] | null;
    };
    readonly Ticket: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly ownerId: CodecTypes['pg/int4@1']['input'] | null;
      readonly subject: CodecTypes['pg/text@1']['input'];
    };
    readonly User: {
      readonly accountId: CodecTypes['pg/int4@1']['input'] | null;
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly kind: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly accounts: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly epics: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly scope: CodecTypes['pg/text@1']['output'];
    };
    readonly features: {
      readonly feature_assignee_person_id: CodecTypes['pg/int4@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly priority: CodecTypes['pg/int4@1']['output'];
    };
    readonly people: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly projects_tbl: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly task_comments: {
      readonly body: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly task_id: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly tasks: {
      readonly bug_assignee_person_id: CodecTypes['pg/int4@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly project_id: CodecTypes['pg/int4@1']['output'] | null;
      readonly reporter_id: CodecTypes['pg/int4@1']['output'] | null;
      readonly severity: CodecTypes['pg/text@1']['output'] | null;
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly type: CodecTypes['pg/text@1']['output'];
    };
    readonly tickets: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly owner_id: CodecTypes['pg/int4@1']['output'] | null;
      readonly subject: CodecTypes['pg/text@1']['output'];
    };
    readonly users: {
      readonly account_id: CodecTypes['pg/int4@1']['output'] | null;
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly kind: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly plan: CodecTypes['pg/text@1']['output'] | null;
      readonly role: CodecTypes['pg/text@1']['output'] | null;
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly accounts: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly epics: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly scope: CodecTypes['pg/text@1']['input'];
    };
    readonly features: {
      readonly feature_assignee_person_id: CodecTypes['pg/int4@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly priority: CodecTypes['pg/int4@1']['input'];
    };
    readonly people: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly projects_tbl: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly task_comments: {
      readonly body: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly task_id: CodecTypes['pg/int4@1']['input'] | null;
    };
    readonly tasks: {
      readonly bug_assignee_person_id: CodecTypes['pg/int4@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly project_id: CodecTypes['pg/int4@1']['input'] | null;
      readonly reporter_id: CodecTypes['pg/int4@1']['input'] | null;
      readonly severity: CodecTypes['pg/text@1']['input'] | null;
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly type: CodecTypes['pg/text@1']['input'];
    };
    readonly tickets: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly owner_id: CodecTypes['pg/int4@1']['input'] | null;
      readonly subject: CodecTypes['pg/text@1']['input'];
    };
    readonly users: {
      readonly account_id: CodecTypes['pg/int4@1']['input'] | null;
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly kind: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly plan: CodecTypes['pg/text@1']['input'] | null;
      readonly role: CodecTypes['pg/text@1']['input'] | null;
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;